05-beads/pyproject.toml
[project]
name = "beads"
version = "0.1.0"
description = "demos from 10 minute physics"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pyside6>=6.9.0", "ncca-ngl"]
